+++ test-helper/infrastructure/README.md
@@ -1,1 +1,7 @@
 ### Start ganache instance
+
+Create `.env` file and add this (replace the URL section with the infure url):
+
+```
+WEB3_FORK_URL=<YOUR URL HERE>
+```
